Move PostWriteHandlers to the IO namespace
